BaseLmmMux: Prevent dts/pts from overflowing.
After 130 seconds of playback, 32-bit numbers overflow for 30 fps and
90khz clock. So we should use 64-bit numbers calculations.

+++ lmm/baselmmmux.cpp
@@ -256,7 +256,7 @@ RawBuffer BaseLmmMux::nextBuffer()
 			pckt.data = (uint8_t *)buf.constData();
 			pckt.size = buf.size();
 			AVRational avrat = context->streams[0]->time_base;
-			pckt.pts = pckt.dts = buf.streamBufferNo() * avrat.den/ avrat.num
+			pckt.pts = pckt.dts = (quint64)buf.streamBufferNo() * avrat.den / avrat.num
 					 / buf.getBufferParameter("fps").toFloat();
 			mInfo("writing next frame");
 			av_write_frame(context, &pckt);
